コンポーネント名の変更
